Refactor entrypoint binary a bit in preparation for logging
This PR splits the entrypoint bin in to separate files for the Runner,
Waiter and PostWriter types.

This is done because we're going to need to modify the Runner while
adding logging and I want to merge the refactoring noise separate from
the logic changes.
