+++ cmd/entrypoint/runner.go
@@ -1,0 +1,32 @@
+package main
+
+import (
+	"os"
+	"os/exec"
+
+	"github.com/tektoncd/pipeline/pkg/entrypoint"
+)
+
+// TODO(jasonhall): Test that original exit code is propagated and that
+// stdout/stderr are collected -- needs e2e tests.
+
+// RealRunner actually runs commands.
+type RealRunner struct{}
+
+var _ entrypoint.Runner = (*RealRunner)(nil)
+
+func (*RealRunner) Run(args ...string) error {
+	if len(args) == 0 {
+		return nil
+	}
+	name, args := args[0], args[1:]
+
+	cmd := exec.Command(name, args...)
+	cmd.Stdout = os.Stdout
+	cmd.Stderr = os.Stderr
+
+	if err := cmd.Run(); err != nil {
+		return err
+	}
+	return nil
+}
